feat: add app.py word_count.html and others files
app.py file: This is the application, written in Python and Flask. It's
an app that allows you to upload a file and count lines, words,
characters, and bytes, similar to the `wc` command on Unix-like systems.

templates folder:
word_count_index.html file: This is the HTML for the client's GET method

word_count.html file: This is the HTML for the client's POST method.

test folder:
test.txt file: This is a test file for internal server-side testing of
the app.