feat(app) add index modify app.py and other change
The app.py file was modified: There are now 3 routes.
1- The root route /, so that a frontend on the same domain can obtain
the HTML in both GET and POST requests. In both cases, it's the same
file, but the results are updated in the POST request. This route is
left open in case a frontend on another domain redirects to the backend.
2- The embedded route sends a fragment of an updated HTML file with the
data, which the frontend must embed.
3- The api route sends a JSON file with the data, and the frontend must
integrate it into the HTML.

The following files were modified in the template folder:
word_count_home.html: This file is for SSR (Single-Screen Request). When
the frontend on your domain makes a GET or POST request, the data is
updated in the POST request.
word_count_embedded.html: This file is a fragment of HTML code with the
updated results.

The index.html file was added. This file contains the ability to receive
a text file and send it to the backend, receive JSON from the backend,
and use the JSON data to modify the DOM to update the results.
It also has the capability to retrieve an HTML fragment from the backend
to embed in the DOM and update the results.