feat(app) modify app.py index.html main.js
app.py:
1- Full types were added, and passed the strict mode mypy and strict
mode pylance tests.
2- Functionality was added to only count lines and bytes when the files
are binary files not encoded with UTF-8, thanks to a new checking
function. This function checks the first 2KB of the file to verify if
it is truly a fully UTF-8 encoded binary file or if it is only the
header and part of the body.
3- The ability to send the full filename was added.
4- More tests were added to all functions using doctest.
5- Complete documentation was added to all functions.
6- The `tuple` object was changed to `NamedTuple` for better data
management.
7- The logic of what each route returns was improved, making everything
more robust and consistent.
8- I tested converting everything to binary with Nuitka, and it worked
perfectly.

index.html: The full filename was added.

main.js: Using filestorage was causing problems with the web app's
caching, so I decided to simplify things and use a regular JSON variable
and make the DOM changes manually.

templates: The option for the full filename was added for HTML templates
for SSR or embedded applications.

test: An image of the Golden State Bridge was added to test the
non-UTF-8 file mode, which only reads lines and bytes.