feat(main) modify main.js app.py add style.css
main.js: Functionality was added to generate an event based on the mode
selection (API or Embedded). This selection automates route selection.
In the case of API, the DOM is updated manually using the returned
JSON data. In the case of Embedded, the DOM is updated manually using
the HTML returned by the server and embedded within a container.

style.css: A stylesheet was added.

templates: The templates were modified to include mode information:
fixed SSR when there is no cross-link, and CSR using an API, or
CSR using an embedded element.

index.html: Functionality was added to choose between API or Embedded
mode.

app.py: Modifications were made to include the mode in the app.