feat(app) modify app.py create wsgi.py
app.py: CORS authentication is now used with a decorator, and only from
the specified routes and domains. Credentials are no longer accepted.

requirements.txt: A file was created with the necessary dependencies and
their versions for automatic installation with pip.

scripts/, styles/: Folders were created to hold the .js and .css files
separately.

wsgi.py: A file was created with the necessary server configuration.